requirements.txt
flask>=3.0
pymysql>=1.1
python-dotenv>=1.0
openpyxl>=3.1
